requirements.txt
Django==3.2.5
wheel
djangorestframework
django-filter
# swagger
drf-yasg
djangorestframework-simplejwt
python-decouple
Pillow
psycopg2-binary
python-slugify
celery
redis